+++ wazo_confd/plugins/endpoint_sip/schema.py
@@ -3,11 +3,14 @@
 
 import re
 import string
+import logging
 
-from marshmallow import fields
+from marshmallow import fields, post_load
 from marshmallow.validate import Length, Regexp, OneOf
 
 from wazo_confd.helpers.mallow import BaseSchema, Link, ListLink
+
+logger = logging.getLogger(__name__)
 
 USERNAME_REGEX = r"^[a-zA-Z0-9_-]{1,40}$"
 SECRET_REGEX = r"^[{}]{{1,80}}$".format(re.escape(string.printable))
@@ -16,7 +19,8 @@ SECRET_REGEX = r"^[{}]{{1,80}}$".format(re.escape(string.printable))
 class SipSchema(BaseSchema):
     id = fields.Integer(dump_only=True)
     tenant_uuid = fields.String(dump_only=True)
-    username = fields.String(validate=Regexp(USERNAME_REGEX), attribute='name')
+    username = fields.String(validate=Regexp(USERNAME_REGEX))
+    name = fields.String(validate=Regexp(USERNAME_REGEX))
     secret = fields.String(validate=Regexp(SECRET_REGEX))
     type = fields.String(validate=OneOf(['friend', 'peer', 'user']))
     host = fields.String(validate=Length(max=255))
@@ -25,6 +29,20 @@ class SipSchema(BaseSchema):
 
     trunk = fields.Nested('TrunkSchema', only=['id', 'links'], dump_only=True)
     line = fields.Nested('LineSchema', only=['id', 'links'], dump_only=True)
+
+    # The set_name_to_username_if_missing method is a compatibility method that
+    # was added in 19.15 to avoid breaking the API. In the old version, the name
+    # could not be specified in the API the username was always copied.
+    @post_load
+    def set_name_to_username_if_missing(self, data, **kwargs):
+        name = data.get('name')
+        if not name and 'username' in data:
+            logger.warning(
+                'DEPRECATION: creating a SIP endpoint with a "username" and no "name" is'
+                ' deprecated. Populate the name field if it is required'
+            )
+            data['name'] = data['username']
+        return data
 
 
 class SipSchemaNullable(SipSchema):