+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+19.15
+-----
+
+* The POST `/1.1/endpoints/sip` now accepts a `name` and a `username` field. If
+  you relied on the behavior that copied to `username` to the `name` field you must
+  upgrade your software before that compatibility layer gets removed in a future version.
+
+
 19.13
 -----